Create and style menu for home page
Styled rectangle areas to look like a menu. Next step is to make it so
the placeholder areas are the same size as the 'projects' selection
area. They appear to be the same size now because 'testing.' takes up
the same number of characters as 'Projects'. If the number of characters
differ then the selection area changes too