Fix top of page styling to honor the box model
Organization of elements was messy with all element containers extending
across the screen. Cleaned up top portion of page but still need to
clean up the about section. Next step is to clean up about section, re-
position menu, and add a portrait image of myself in the empty space
at the top right of the page